Fix smartthingcluster import
